pyproject.toml
[project]
name = "api-gateway"
version = "0.1.0"
description = ""
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
    "pyjwt>=2.9.0",
    "pydantic-settings>=2.7.0",
]

[tool.uv]
dev-dependencies = []
